Add Docker

+++ db/query/account.sql
@@ -26,8 +26,8 @@ LIMIT $2
 OFFSET $3;
 
 -- name: UpdateAccount :one
-UPDATE accounts SET balance = $2
-WHERE id = $1
+UPDATE accounts SET balance = $3
+WHERE id = $1 AND owner = $2
 RETURNING *;
 
 -- name: AddAccountBalance :one
@@ -36,4 +36,4 @@ WHERE id = sqlc.arg(id)
 RETURNING *;
 
 -- name: DeleteAccount :exec
-DELETE FROM accounts WHERE id = $1;
+DELETE FROM accounts WHERE id = $1 AND owner = $2;
